PoC use the Centered chrome in the Landing
