+++ templates/chromes/Centered/styles.scss
@@ -1,0 +1,30 @@
+@import '../../../css/settings';
+@import '../../../css/mixins/index';
+
+.chrome--centered {
+  text-align: center;
+}
+
+.chrome--centered__title {
+  padding-top: ($grid * 7.2);
+}
+
+.chrome--centered__title + .chrome--centered__paragraph--primary {
+  padding-top: ($grid * 2.4);
+}
+
+.chrome--centered__button--accept {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: ($grid * 3);
+  text-align: center;
+}
+
+.chrome--centered__paragraph--primary + .chrome--centered__button--accept {
+  margin-top: ($grid * 8.2);
+}
+
+.chrome--centered__paragraph--legal {
+  padding-top: ($grid * 6.2);
+}
